+++ backend/src/main/kotlin/com/infraView/alarm/database/SpringDataAlarmRepository.kt
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface SpringDataAlarmRepository : JpaRepository<AlarmJpaEntity, Int>
+interface SpringDataAlarmRepository : JpaRepository<AlarmJpaEntity, Long>
