+++ backend/src/main/kotlin/com/infraView/incident/domain/IncidentPort.kt
@@ -2,8 +2,8 @@ package com.infraView.incident.domain
 
 
 interface IncidentPort {
-    fun getById(id: Int): Incident?
+    fun getById(id: Long): Incident?
     fun getAll(): List<Incident>
     fun save(incident: Incident): Incident
-    fun deleteById(id: Int)
+    fun deleteById(id: Long)
 }
