+++ backend/src/main/kotlin/com/infraView/incident/domain/ManageIncidentUseCase.kt
@@ -3,8 +3,8 @@ package com.infraView.incident.domain
 
 interface ManageIncidentUseCase {
     fun getAll(): List<Incident>
-    fun getById(id: Int): Incident?
+    fun getById(id: Long): Incident?
     fun save(incident: Incident): Incident
-    fun endIncident(incidentId: Int): Incident?
-    fun delete(incidentId: Int)
+    fun endIncident(incidentId: Long): Incident?
+    fun delete(incidentId: Long)
 }
