+++ backend/src/main/kotlin/com/infraView/video/domain/ManageVideoRecordingUseCase.kt
@@ -4,9 +4,9 @@ import java.time.OffsetDateTime
 
 interface ManageVideoRecordingUseCase {
     fun getAll(): List<VideoRecording>
-    fun getById(id: Int): VideoRecording?
-    fun getByIncidentId(incidentId: Int): List<VideoRecording>
-    fun startRecording(incidentId: Int, startedAt: OffsetDateTime): VideoRecording
-    fun updateRecording(id: Int, endedAt: OffsetDateTime?, storageKey: String?, fileSizeBytes: Long?, durationSec: Int?): VideoRecording?
-    fun delete(id: Int)
+    fun getById(id: Long): VideoRecording?
+    fun getByIncidentId(incidentId: Long): List<VideoRecording>
+    fun startRecording(incidentId: Long, startedAt: OffsetDateTime): VideoRecording
+    fun updateRecording(id: Long, endedAt: OffsetDateTime?, storageKey: String?, fileSizeBytes: Long?, durationSec: Int?): VideoRecording?
+    fun delete(id: Long)
 }
